set alarms color in internments items

+++ app/src/main/java/com/mapx/kosten/mosimpa/domain/entites/AlarmsEntity.kt
@@ -1,10 +1,10 @@
 package com.mapx.kosten.mosimpa.domain.entites
 
 data class AlarmsEntity (
-    val spo2_lt: Int = 0,
-    val hr_lt: Int = 0,
-    val hr_gt: Int = 0,
-    val bt_gt: Int = 0,
-    val bp_sys_lt: Int = 0,
-    val bp_sys_gt: Int = 0
+    val spo2_lt: Float = 0F,
+    val hr_lt: Float = 0F,
+    val hr_gt: Float = 0F,
+    val bt_gt: Float = 0F,
+    val bp_sys_lt: Float = 0F,
+    val bp_sys_gt: Float = 0F
 )
